extract `resolveOrigin`
