+++ src/shapes/Object/ObjectOrigin.ts
@@ -5,27 +5,7 @@ import { transformPoint } from '../../util/misc/matrix';
 import { sizeAfterTransform } from '../../util/misc/objectTransforms';
 import { degreesToRadians } from '../../util/misc/radiansDegreesConversion';
 import { CommonMethods } from '../../mixins/shared_methods.mixin';
-
-const originOffset = {
-  left: -0.5,
-  top: -0.5,
-  center: 0,
-  bottom: 0.5,
-  right: 0.5,
-};
-
-/**
- * Resolves origin value relative to center
- * @private
- * @param {TOriginX | TOriginY} originValue originX / originY
- * @returns number
- */
-export const resolveOrigin = (
-  originValue: TOriginX | TOriginY | number
-): number =>
-  typeof originValue === 'string'
-    ? originOffset[originValue]
-    : originValue - 0.5;
+import { resolveOrigin } from '../../util/misc/resolveOrigin';
 
 export class ObjectOrigin<EventSpec> extends CommonMethods<EventSpec> {
   /**
